Added Saving and Loading for Combat Manager and Combat Handlers
-Changed FactionManagerScript to use int comparsion for enums
-Changed faction comparision in CharacterInfo to int comparision of the enums
-Added SetStoppingDistance() for NPCInfo for loading inCombat
-Added check in CombatHandler for if it was loaded to prevent duplicate loading of information
-Added NPCs and Player as children of the NPCManager to better keep track of them and prevent loading all resources

+++ Assets/Scripts/Game Scripts/FactionManagerScript.cs
@@ -78,7 +78,7 @@ public class FactionManagerScript : MonoBehaviour {
     /// <param name="b"></param>
     public void CreateFactionRelation(Faction a, Faction b) {
         
-        if (a != b) {
+        if ((int)a != (int)b) {
             if (!DoesFactionRelationExist(a, b)) {
                 FactionRelation fr = new FactionRelation();
                 fr.faction1 = a;
@@ -94,11 +94,11 @@ public class FactionManagerScript : MonoBehaviour {
     /// <param name="b"></param>
     /// <param name="newRelation"></param>
     public void SetFactionRelation(Faction a,Faction b, float newRelation) {
-        if (a == b) {
+        if ((int)a == (int)b) {
             return;
         }
         foreach (FactionRelation fr in factionRelations) {
-            if ((a == fr.faction1 || a == fr.faction2) && (b == fr.faction1 || b == fr.faction2)) {
+            if (((int)a == (int)fr.faction1 || (int)a == (int)fr.faction2) && ((int)b == (int)fr.faction1 || (int)b == (int)fr.faction2)) {
                 fr.opinion = newRelation;
                 Validate(fr);
                 return;
@@ -126,11 +126,11 @@ public class FactionManagerScript : MonoBehaviour {
     /// <param name="b"></param>
     /// <returns></returns>
     public float GetFactionRelation(Faction a,Faction b) {
-        if (a == b) {
+        if ((int)a == (int)b) {
             return 0;
         }
         foreach (FactionRelation fr in factionRelations) {
-            if ((a == fr.faction1 || a == fr.faction2) && (b == fr.faction1 || b == fr.faction2)) {
+            if (((int)a == (int)fr.faction1 || (int)a == (int)fr.faction2) && ((int)b == (int)fr.faction1 || (int)b == (int)fr.faction2)) {
                 return fr.opinion;
             }
         }
@@ -145,11 +145,11 @@ public class FactionManagerScript : MonoBehaviour {
     /// <param name="addedValue"></param>
     /// <returns></returns>
     public float AddToFactionRelation(Faction a, Faction b, float addedValue) { 
-        if (a == b) {
+        if ((int)a == (int)b) {
             return 0;
         }
         foreach (FactionRelation fr in factionRelations) {
-            if ((a == fr.faction1 || a == fr.faction2) && (b == fr.faction1 || b == fr.faction2)) {
+            if (((int)a == (int)fr.faction1 || (int)a == (int)fr.faction2) && ((int)b == (int)fr.faction1 || (int)b == (int)fr.faction2)) {
                 return (fr.opinion += addedValue);
             }
         }
@@ -166,11 +166,11 @@ public class FactionManagerScript : MonoBehaviour {
     /// <param name="b"></param>
     /// <returns></returns>
     public bool DoesHate(Faction a, Faction b) {
-        if (a == b) {
+        if ((int)a == (int)b) {
             return false;
         }
         foreach (FactionRelation fr in factionRelations) {
-            if ((a == fr.faction1 || a == fr.faction2) && (b == fr.faction1 || b == fr.faction2)) {
+            if (((int)a == (int)fr.faction1 || (int)a == (int)fr.faction2) && ((int)b == (int)fr.faction1 || (int)b == (int)fr.faction2)) {
                 return (fr.opinion < doesHateUpperBound);
             }
         }
@@ -184,11 +184,11 @@ public class FactionManagerScript : MonoBehaviour {
     /// <param name="b"></param>
     /// <returns></returns>
     public bool DoesDislike(Faction a, Faction b) {
-        if (a == b) {
+        if ((int)a == (int)b) {
             return false;
         }
         foreach (FactionRelation fr in factionRelations) {
-            if ((a == fr.faction1 || a == fr.faction2) && (b == fr.faction1 || b == fr.faction2)) { 
+            if (((int)a == (int)fr.faction1 || (int)a == (int)fr.faction2) && ((int)b == (int)fr.faction1 || (int)b == (int)fr.faction2)) { 
                 return (fr.opinion < doesDislikeUpperBound);
             }
         }
@@ -202,11 +202,11 @@ public class FactionManagerScript : MonoBehaviour {
     /// <param name="b"></param>
     /// <returns></returns>
     public bool IsNeutral(Faction a, Faction b) {
-        if (a == b) {
+        if ((int)a == (int)b) {
             return false;
         }
         foreach (FactionRelation fr in factionRelations) {
-            if ((a == fr.faction1 || a == fr.faction2) && (b == fr.faction1 || b == fr.faction2)) {
+            if (((int)a == (int)fr.faction1 || (int)a == (int)fr.faction2) && ((int)b == (int)fr.faction1 || (int)b == (int)fr.faction2)) {
                 return ((fr.opinion >= isNeutralLowerBound) && (fr.opinion <= isNeutralUpperBound));
             }
         }
@@ -220,11 +220,11 @@ public class FactionManagerScript : MonoBehaviour {
     /// <param name="b"></param>
     /// <returns></returns>
     public bool DoesLike(Faction a, Faction b) {
-        if (a == b) {
+        if ((int)a == (int)b) {
             return false;
         }
         foreach (FactionRelation fr in factionRelations) {
-            if ((a == fr.faction1 || a == fr.faction2) && (b == fr.faction1 || b == fr.faction2)) {
+            if (((int)a == (int)fr.faction1 || (int)a == (int)fr.faction2) && ((int)b == (int)fr.faction1 || (int)b == (int)fr.faction2)) {
                 return (fr.opinion > doesLikeLowerBound);
             }
         }
@@ -242,7 +242,7 @@ public class FactionManagerScript : MonoBehaviour {
             return false;
         }
         foreach (FactionRelation fr in factionRelations) {
-            if ((a == fr.faction1 || a == fr.faction2) && (b == fr.faction1 || b == fr.faction2)) {
+            if (((int)a == (int)fr.faction1 || (int)a == (int)fr.faction2) && ((int)b == (int)fr.faction1 || (int)b == (int)fr.faction2)) {
                 return (fr.opinion > doesLoveLowerBound);
             }
         }
